chore(entries-dtos): rename
